requirements.txt
pandas>=2.2.0
numpy>=1.26.0
yfinance>=1.3.0
statsmodels>=0.14.5
scikit-learn>=1.4.0
xgboost>=2.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
openai>=1.109.0
